Module google_dfareporting3d4::api

Structs

Account
Contains properties of a Campaign Manager account.

AccountActiveAdSummary
Gets a summary of active ads in an account.

AccountActiveAdSummaryGetCall
Gets the account's active ad summary by account ID.

AccountActiveAdSummaryMethods
A builder providing access to all methods supported on accountActiveAdSummary resources. It is not used directly, but through the Dfareporting hub.

AccountGetCall
Gets one account by ID.

AccountListCall
Retrieves the list of accounts, possibly filtered. This method supports paging.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the Dfareporting hub.

AccountPatchCall
Updates an existing account. This method supports patch semantics.

AccountPermission
AccountPermissions contains information about a particular account permission. Some features of Campaign Manager require an account permission to be present in the account.

AccountPermissionGetCall
Gets one account permission by ID.

AccountPermissionGroup
AccountPermissionGroups contains a mapping of permission group IDs to names. A permission group is a grouping of account permissions.

AccountPermissionGroupGetCall
Gets one account permission group by ID.

AccountPermissionGroupListCall
Retrieves the list of account permission groups.

AccountPermissionGroupMethods
A builder providing access to all methods supported on accountPermissionGroup resources. It is not used directly, but through the Dfareporting hub.

AccountPermissionGroupsListResponse
Account Permission Group List Response

AccountPermissionListCall
Retrieves the list of account permissions.

AccountPermissionMethods
A builder providing access to all methods supported on accountPermission resources. It is not used directly, but through the Dfareporting hub.

AccountPermissionsListResponse
Account Permission List Response

AccountUpdateCall
Updates an existing account.

AccountUserProfile
AccountUserProfiles contains properties of a Campaign Manager user profile. This resource is specifically for managing user profiles, whereas UserProfiles is for accessing the API.

AccountUserProfileGetCall
Gets one account user profile by ID.

AccountUserProfileInsertCall
Inserts a new account user profile.

AccountUserProfileListCall
Retrieves a list of account user profiles, possibly filtered. This method supports paging.

AccountUserProfileMethods
A builder providing access to all methods supported on accountUserProfile resources. It is not used directly, but through the Dfareporting hub.

AccountUserProfilePatchCall
Updates an existing account user profile. This method supports patch semantics.

AccountUserProfileUpdateCall
Updates an existing account user profile.

AccountUserProfilesListResponse
Account User Profile List Response

AccountsListResponse
Account List Response

Activities
Represents an activity group.

Ad
Contains properties of a Campaign Manager ad.

AdBlockingConfiguration
Campaign ad blocking settings.

AdGetCall
Gets one ad by ID.

AdInsertCall
Inserts a new ad.

AdListCall
Retrieves a list of ads, possibly filtered. This method supports paging.

AdMethods
A builder providing access to all methods supported on ad resources. It is not used directly, but through the Dfareporting hub.

AdPatchCall
Updates an existing ad. This method supports patch semantics.

AdSlot
Ad Slot

AdUpdateCall
Updates an existing ad.

AdsListResponse
Ad List Response

Advertiser
Contains properties of a Campaign Manager advertiser.

AdvertiserGetCall
Gets one advertiser by ID.

AdvertiserGroup
Groups advertisers together so that reports can be generated for the entire group at once.

AdvertiserGroupDeleteCall
Deletes an existing advertiser group.

AdvertiserGroupGetCall
Gets one advertiser group by ID.

AdvertiserGroupInsertCall
Inserts a new advertiser group.

AdvertiserGroupListCall
Retrieves a list of advertiser groups, possibly filtered. This method supports paging.

AdvertiserGroupMethods
A builder providing access to all methods supported on advertiserGroup resources. It is not used directly, but through the Dfareporting hub.

AdvertiserGroupPatchCall
Updates an existing advertiser group. This method supports patch semantics.

AdvertiserGroupUpdateCall
Updates an existing advertiser group.

AdvertiserGroupsListResponse
Advertiser Group List Response

AdvertiserInsertCall
Inserts a new advertiser.

AdvertiserLandingPageGetCall
Gets one landing page by ID.

AdvertiserLandingPageInsertCall
Inserts a new landing page.

AdvertiserLandingPageListCall
Retrieves a list of landing pages.

AdvertiserLandingPageMethods
A builder providing access to all methods supported on advertiserLandingPage resources. It is not used directly, but through the Dfareporting hub.

AdvertiserLandingPagePatchCall
Updates an existing advertiser landing page. This method supports patch semantics.

AdvertiserLandingPageUpdateCall
Updates an existing landing page.

AdvertiserLandingPagesListResponse
Landing Page List Response

AdvertiserListCall
Retrieves a list of advertisers, possibly filtered. This method supports paging.

AdvertiserMethods
A builder providing access to all methods supported on advertiser resources. It is not used directly, but through the Dfareporting hub.

AdvertiserPatchCall
Updates an existing advertiser. This method supports patch semantics.

AdvertiserUpdateCall
Updates an existing advertiser.

AdvertisersListResponse
Advertiser List Response

AudienceSegment
Audience Segment.

AudienceSegmentGroup
Audience Segment Group.

Browser
Contains information about a browser that can be targeted by ads.

BrowserListCall
Retrieves a list of browsers.

BrowserMethods
A builder providing access to all methods supported on browser resources. It is not used directly, but through the Dfareporting hub.

BrowsersListResponse
Browser List Response

Campaign
Contains properties of a Campaign Manager campaign.

CampaignCreativeAssociation
Identifies a creative which has been associated with a given campaign.

CampaignCreativeAssociationInsertCall
Associates a creative with the specified campaign. This method creates a default ad with dimensions matching the creative in the campaign if such a default ad does not exist already.

CampaignCreativeAssociationListCall
Retrieves the list of creative IDs associated with the specified campaign. This method supports paging.

CampaignCreativeAssociationMethods
A builder providing access to all methods supported on campaignCreativeAssociation resources. It is not used directly, but through the Dfareporting hub.

CampaignCreativeAssociationsListResponse
Campaign Creative Association List Response

CampaignGetCall
Gets one campaign by ID.

CampaignInsertCall
Inserts a new campaign.

CampaignListCall
Retrieves a list of campaigns, possibly filtered. This method supports paging.

CampaignManagerIds
Campaign Manager IDs related to the custom event.

CampaignMethods
A builder providing access to all methods supported on campaign resources. It is not used directly, but through the Dfareporting hub.

CampaignPatchCall
Updates an existing campaign. This method supports patch semantics.

CampaignUpdateCall
Updates an existing campaign.

CampaignsListResponse
Campaign List Response

ChangeLog
Describes a change that a user has made to a resource.

ChangeLogGetCall
Gets one change log by ID.

ChangeLogListCall
Retrieves a list of change logs. This method supports paging.

ChangeLogMethods
A builder providing access to all methods supported on changeLog resources. It is not used directly, but through the Dfareporting hub.

ChangeLogsListResponse
Change Log List Response

ChannelGrouping
Represents a DfaReporting channel grouping.

ChannelGroupingRule
Represents a DfaReporting channel grouping rule.

CitiesListResponse
City List Response

City
Contains information about a city that can be targeted by ads.

CityListCall
Retrieves a list of cities, possibly filtered.

CityMethods
A builder providing access to all methods supported on city resources. It is not used directly, but through the Dfareporting hub.

ClickTag
Creative Click Tag.

ClickThroughUrl
Click-through URL

ClickThroughUrlSuffixProperties
Click Through URL Suffix settings.

CompanionClickThroughOverride
Companion Click-through override.

CompanionSetting
Companion Settings

CompatibleFields
Represents a response to the queryCompatibleFields method.

ConnectionType
Contains information about an internet connection type that can be targeted by ads. Clients can use the connection type to target mobile vs. broadband users.

ConnectionTypeGetCall
Gets one connection type by ID.

ConnectionTypeListCall
Retrieves a list of connection types.

ConnectionTypeMethods
A builder providing access to all methods supported on connectionType resources. It is not used directly, but through the Dfareporting hub.

ConnectionTypesListResponse
Connection Type List Response

ContentCategoriesListResponse
Content Category List Response

ContentCategory
Organizes placements according to the contents of their associated webpages.

ContentCategoryDeleteCall
Deletes an existing content category.

ContentCategoryGetCall
Gets one content category by ID.

ContentCategoryInsertCall
Inserts a new content category.

ContentCategoryListCall
Retrieves a list of content categories, possibly filtered. This method supports paging.

ContentCategoryMethods
A builder providing access to all methods supported on contentCategory resources. It is not used directly, but through the Dfareporting hub.

ContentCategoryPatchCall
Updates an existing content category. This method supports patch semantics.

ContentCategoryUpdateCall
Updates an existing content category.

Conversion
A Conversion represents when a user successfully performs a desired action after seeing an ad.

ConversionBatchinsertCall
Inserts conversions.

ConversionBatchupdateCall
Updates existing conversions.

ConversionError
The error code and description for a conversion that failed to insert or update.

ConversionMethods
A builder providing access to all methods supported on conversion resources. It is not used directly, but through the Dfareporting hub.

ConversionStatus
The original conversion that was inserted or updated and whether there were any errors.

ConversionsBatchInsertRequest
Insert Conversions Request.

ConversionsBatchInsertResponse
Insert Conversions Response.

ConversionsBatchUpdateRequest
Update Conversions Request.

ConversionsBatchUpdateResponse
Update Conversions Response.

CountriesListResponse
Country List Response

Country
Contains information about a country that can be targeted by ads.

CountryGetCall
Gets one country by ID.

CountryListCall
Retrieves a list of countries.

CountryMethods
A builder providing access to all methods supported on country resources. It is not used directly, but through the Dfareporting hub.

Creative
Contains properties of a Creative.

CreativeAsset
Creative Asset.

CreativeAssetId
Creative Asset ID.

CreativeAssetInsertCall
Inserts a new creative asset.

CreativeAssetMetadata
CreativeAssets contains properties of a creative asset file which will be uploaded or has already been uploaded. Refer to the creative sample code for how to upload assets and insert a creative.

CreativeAssetMethods
A builder providing access to all methods supported on creativeAsset resources. It is not used directly, but through the Dfareporting hub.

CreativeAssetSelection
Encapsulates the list of rules for asset selection and a default asset in case none of the rules match. Applicable to INSTREAM_VIDEO creatives.

CreativeAssignment
Creative Assignment.

CreativeClickThroughUrl
Click-through URL

CreativeCustomEvent
Creative Custom Event.

CreativeField
Contains properties of a creative field.

CreativeFieldAssignment
Creative Field Assignment.

CreativeFieldDeleteCall
Deletes an existing creative field.

CreativeFieldGetCall
Gets one creative field by ID.

CreativeFieldInsertCall
Inserts a new creative field.

CreativeFieldListCall
Retrieves a list of creative fields, possibly filtered. This method supports paging.

CreativeFieldMethods
A builder providing access to all methods supported on creativeField resources. It is not used directly, but through the Dfareporting hub.

CreativeFieldPatchCall
Updates an existing creative field. This method supports patch semantics.

CreativeFieldUpdateCall
Updates an existing creative field.

CreativeFieldValue
Contains properties of a creative field value.

CreativeFieldValueDeleteCall
Deletes an existing creative field value.

CreativeFieldValueGetCall
Gets one creative field value by ID.

CreativeFieldValueInsertCall
Inserts a new creative field value.

CreativeFieldValueListCall
Retrieves a list of creative field values, possibly filtered. This method supports paging.

CreativeFieldValueMethods
A builder providing access to all methods supported on creativeFieldValue resources. It is not used directly, but through the Dfareporting hub.

CreativeFieldValuePatchCall
Updates an existing creative field value. This method supports patch semantics.

CreativeFieldValueUpdateCall
Updates an existing creative field value.

CreativeFieldValuesListResponse
Creative Field Value List Response

CreativeFieldsListResponse
Creative Field List Response

CreativeGetCall
Gets one creative by ID.

CreativeGroup
Contains properties of a creative group.

CreativeGroupAssignment
Creative Group Assignment.

CreativeGroupGetCall
Gets one creative group by ID.

CreativeGroupInsertCall
Inserts a new creative group.

CreativeGroupListCall
Retrieves a list of creative groups, possibly filtered. This method supports paging.

CreativeGroupMethods
A builder providing access to all methods supported on creativeGroup resources. It is not used directly, but through the Dfareporting hub.

CreativeGroupPatchCall
Updates an existing creative group. This method supports patch semantics.

CreativeGroupUpdateCall
Updates an existing creative group.

CreativeGroupsListResponse
Creative Group List Response

CreativeInsertCall
Inserts a new creative.

CreativeListCall
Retrieves a list of creatives, possibly filtered. This method supports paging.

CreativeMethods
A builder providing access to all methods supported on creative resources. It is not used directly, but through the Dfareporting hub.

CreativeOptimizationConfiguration
Creative optimization settings.

CreativePatchCall
Updates an existing creative. This method supports patch semantics.

CreativeRotation
Creative Rotation.

CreativeUpdateCall
Updates an existing creative.

CreativesListResponse
Creative List Response

CrossDimensionReachReportCompatibleFields
Represents fields that are compatible to be selected for a report of type "CROSS_DIMENSION_REACH".

CustomEvent
Experimental feature (no support provided) A custom event represents a third party impression, a third party click, an annotation on a first party impression, or an annotation on a first party click.

CustomEventBatchinsertCall
Inserts custom events.

CustomEventClickAnnotation
Annotate a click event.

CustomEventError
The error code and description for a custom event that failed to insert.

CustomEventImpressionAnnotation
Annotate an impression.

CustomEventInsert
Custom event to be inserted.

CustomEventMethods
A builder providing access to all methods supported on customEvent resources. It is not used directly, but through the Dfareporting hub.

CustomEventStatus
The original custom event that was inserted and whether there were any errors.

CustomEventsBatchInsertRequest
Insert Custom Events Request.

CustomEventsBatchInsertResponse
Insert Custom Events Response.

CustomFloodlightVariable
A custom floodlight variable.

CustomRichMediaEvents
Represents a Custom Rich Media Events group.

CustomVariable
Custom variable.

CustomViewabilityMetric
Custom Viewability Metric

CustomViewabilityMetricConfiguration
The attributes, like playtime and percent onscreen, that define the Custom Viewability Metric.

DV3Ids
DV360 IDs related to the custom event.

DateRange
Represents a date range.

DayPartTargeting
Day Part Targeting.

DeepLink
Contains information about a landing page deep link.

DefaultClickThroughEventTagProperties
Properties of inheriting and overriding the default click-through event tag. A campaign may override the event tag defined at the advertiser level, and an ad may also override the campaign's setting further.

DeliverySchedule
Delivery Schedule.

Dfareporting
Central instance to access all Dfareporting related resource activities

DfpSettings
Google Ad Manager Settings

Dimension
Represents a dimension.

DimensionFilter
Represents a dimension filter.

DimensionValue
Represents a DimensionValue resource.

DimensionValueList
Represents the list of DimensionValue resources.

DimensionValueMethods
A builder providing access to all methods supported on dimensionValue resources. It is not used directly, but through the Dfareporting hub.

DimensionValueQueryCall
Retrieves list of report dimension values for a list of filters.

DimensionValueRequest
Represents a DimensionValuesRequest.

DirectorySite
DirectorySites contains properties of a website from the Site Directory. Sites need to be added to an account via the Sites resource before they can be assigned to a placement.

DirectorySiteGetCall
Gets one directory site by ID.

DirectorySiteInsertCall
Inserts a new directory site.

DirectorySiteListCall
Retrieves a list of directory sites, possibly filtered. This method supports paging.

DirectorySiteMethods
A builder providing access to all methods supported on directorySite resources. It is not used directly, but through the Dfareporting hub.

DirectorySiteSettings
Directory Site Settings

DirectorySitesListResponse
Directory Site List Response

DisjunctiveMatchStatement
Represents a Disjunctive Match Statement resource, which is a conjunction (and) of disjunctive (or) boolean statements.

DynamicTargetingKey
Contains properties of a dynamic targeting key. Dynamic targeting keys are unique, user-friendly labels, created at the advertiser level in DCM, that can be assigned to ads, creatives, and placements and used for targeting with Studio dynamic creatives. Use these labels instead of numeric Campaign Manager IDs (such as placement IDs) to save time and avoid errors in your dynamic feeds.

DynamicTargetingKeyDeleteCall
Deletes an existing dynamic targeting key.

DynamicTargetingKeyInsertCall
Inserts a new dynamic targeting key. Keys must be created at the advertiser level before being assigned to the advertiser's ads, creatives, or placements. There is a maximum of 1000 keys per advertiser, out of which a maximum of 20 keys can be assigned per ad, creative, or placement.

DynamicTargetingKeyListCall
Retrieves a list of dynamic targeting keys.

DynamicTargetingKeyMethods
A builder providing access to all methods supported on dynamicTargetingKey resources. It is not used directly, but through the Dfareporting hub.

DynamicTargetingKeysListResponse
Dynamic Targeting Key List Response

EncryptionInfo
A description of how user IDs are encrypted.

EventFilter
Represents a DfaReporting event filter.

EventTag
Contains properties of an event tag.

EventTagDeleteCall
Deletes an existing event tag.

EventTagGetCall
Gets one event tag by ID.

EventTagInsertCall
Inserts a new event tag.

EventTagListCall
Retrieves a list of event tags, possibly filtered.

EventTagMethods
A builder providing access to all methods supported on eventTag resources. It is not used directly, but through the Dfareporting hub.

EventTagOverride
Event tag override information.

EventTagPatchCall
Updates an existing event tag. This method supports patch semantics.

EventTagUpdateCall
Updates an existing event tag.

EventTagsListResponse
Event Tag List Response

File
Represents a File resource. A file contains the metadata for a report run. It shows the status of the run and holds the URLs to the generated report data if the run is finished and the status is "REPORT_AVAILABLE".

FileGetCall
Retrieves a report file by its report ID and file ID. This method supports media download.

FileList
List of files for a report.

FileListCall
Lists files for a user profile.

FileMethods
A builder providing access to all methods supported on file resources. It is not used directly, but through the Dfareporting hub.

FileUrls
The URLs where the completed report file can be downloaded.

Flight
Flight

FloodlightActivitiesGenerateTagResponse
Floodlight Activity GenerateTag Response

FloodlightActivitiesListResponse
Floodlight Activity List Response

FloodlightActivity
Contains properties of a Floodlight activity.

FloodlightActivityDeleteCall
Deletes an existing floodlight activity.

FloodlightActivityDynamicTag
Dynamic Tag

FloodlightActivityGeneratetagCall
Generates a tag for a floodlight activity.

FloodlightActivityGetCall
Gets one floodlight activity by ID.

FloodlightActivityGroup
Contains properties of a Floodlight activity group.

FloodlightActivityGroupGetCall
Gets one floodlight activity group by ID.

FloodlightActivityGroupInsertCall
Inserts a new floodlight activity group.

FloodlightActivityGroupListCall
Retrieves a list of floodlight activity groups, possibly filtered. This method supports paging.

FloodlightActivityGroupMethods
A builder providing access to all methods supported on floodlightActivityGroup resources. It is not used directly, but through the Dfareporting hub.

FloodlightActivityGroupPatchCall
Updates an existing floodlight activity group. This method supports patch semantics.

FloodlightActivityGroupUpdateCall
Updates an existing floodlight activity group.

FloodlightActivityGroupsListResponse
Floodlight Activity Group List Response

FloodlightActivityInsertCall
Inserts a new floodlight activity.

FloodlightActivityListCall
Retrieves a list of floodlight activities, possibly filtered. This method supports paging.

FloodlightActivityMethods
A builder providing access to all methods supported on floodlightActivity resources. It is not used directly, but through the Dfareporting hub.

FloodlightActivityPatchCall
Updates an existing floodlight activity. This method supports patch semantics.

FloodlightActivityPublisherDynamicTag
Publisher Dynamic Tag

FloodlightActivityUpdateCall
Updates an existing floodlight activity.

FloodlightConfiguration
Contains properties of a Floodlight configuration.

FloodlightConfigurationGetCall
Gets one floodlight configuration by ID.

FloodlightConfigurationListCall
Retrieves a list of floodlight configurations, possibly filtered.

FloodlightConfigurationMethods
A builder providing access to all methods supported on floodlightConfiguration resources. It is not used directly, but through the Dfareporting hub.

FloodlightConfigurationPatchCall
Updates an existing floodlight configuration. This method supports patch semantics.

FloodlightConfigurationUpdateCall
Updates an existing floodlight configuration.

FloodlightConfigurationsListResponse
Floodlight Configuration List Response

FloodlightReportCompatibleFields
Represents fields that are compatible to be selected for a report of type "FlOODLIGHT".

FrequencyCap
Frequency Cap.

FsCommand
FsCommand.

GeoTargeting
Geographical Targeting.

InventoryItem
Represents a buy from the Planning inventory store.

InventoryItemGetCall
Gets one inventory item by ID.

InventoryItemListCall
Retrieves a list of inventory items, possibly filtered. This method supports paging.

InventoryItemMethods
A builder providing access to all methods supported on inventoryItem resources. It is not used directly, but through the Dfareporting hub.

InventoryItemsListResponse
Inventory item List Response

KeyValueTargetingExpression
Key Value Targeting Expression.

LandingPage
Contains information about where a user's browser is taken after the user clicks an ad.

Language
Contains information about a language that can be targeted by ads.

LanguageListCall
Retrieves a list of languages.

LanguageMethods
A builder providing access to all methods supported on language resources. It is not used directly, but through the Dfareporting hub.

LanguageTargeting
Language Targeting.

LanguagesListResponse
Language List Response

LastModifiedInfo
Modification timestamp.

ListPopulationClause
A group clause made up of list population terms representing constraints joined by ORs.

ListPopulationRule
Remarketing List Population Rule.

ListPopulationTerm
Remarketing List Population Rule Term.

ListTargetingExpression
Remarketing List Targeting Expression.

LookbackConfiguration
Lookback configuration settings.

Metric
Represents a metric.

Metro
Contains information about a metro region that can be targeted by ads.

MetroListCall
Retrieves a list of metros.

MetroMethods
A builder providing access to all methods supported on metro resources. It is not used directly, but through the Dfareporting hub.

MetrosListResponse
Metro List Response

MobileApp
Contains information about a mobile app. Used as a landing page deep link.

MobileAppGetCall
Gets one mobile app by ID.

MobileAppListCall
Retrieves list of available mobile apps.

MobileAppMethods
A builder providing access to all methods supported on mobileApp resources. It is not used directly, but through the Dfareporting hub.

MobileAppsListResponse
Mobile app List Response

MobileCarrier
Contains information about a mobile carrier that can be targeted by ads.

MobileCarrierGetCall
Gets one mobile carrier by ID.

MobileCarrierListCall
Retrieves a list of mobile carriers.

MobileCarrierMethods
A builder providing access to all methods supported on mobileCarrier resources. It is not used directly, but through the Dfareporting hub.

MobileCarriersListResponse
Mobile Carrier List Response

ObaIcon
Online Behavioral Advertiser icon.

ObjectFilter
Object Filter.

OffsetPosition
Offset Position.

OmnitureSettings
Omniture Integration Settings.

OperatingSystem
Contains information about an operating system that can be targeted by ads.

OperatingSystemGetCall
Gets one operating system by DART ID.

OperatingSystemListCall
Retrieves a list of operating systems.

OperatingSystemMethods
A builder providing access to all methods supported on operatingSystem resources. It is not used directly, but through the Dfareporting hub.

OperatingSystemVersion
Contains information about a particular version of an operating system that can be targeted by ads.

OperatingSystemVersionGetCall
Gets one operating system version by ID.

OperatingSystemVersionListCall
Retrieves a list of operating system versions.

OperatingSystemVersionMethods
A builder providing access to all methods supported on operatingSystemVersion resources. It is not used directly, but through the Dfareporting hub.

OperatingSystemVersionsListResponse
Operating System Version List Response

OperatingSystemsListResponse
Operating System List Response

OptimizationActivity
Creative optimization activity.

Order
Describes properties of a Planning order.

OrderContact
Contact of an order.

OrderDocument
Contains properties of a Planning order document.

OrderDocumentGetCall
Gets one order document by ID.

OrderDocumentListCall
Retrieves a list of order documents, possibly filtered. This method supports paging.

OrderDocumentMethods
A builder providing access to all methods supported on orderDocument resources. It is not used directly, but through the Dfareporting hub.

OrderDocumentsListResponse
Order document List Response

OrderGetCall
Gets one order by ID.

OrderListCall
Retrieves a list of orders, possibly filtered. This method supports paging.

OrderMethods
A builder providing access to all methods supported on order resources. It is not used directly, but through the Dfareporting hub.

OrdersListResponse
Order List Response

PathFilter
Represents a DfaReporting path filter.

PathReportCompatibleFields
Represents fields that are compatible to be selected for a report of type "PATH".

PathReportDimensionValue
Represents a PathReportDimensionValue resource.

PathToConversionReportCompatibleFields
Represents fields that are compatible to be selected for a report of type "PATH_TO_CONVERSION".

Placement
Contains properties of a placement.

PlacementAssignment
Placement Assignment.

PlacementGeneratetagCall
Generates tags for a placement.

PlacementGetCall
Gets one placement by ID.

PlacementGroup
Contains properties of a package or roadblock.

PlacementGroupGetCall
Gets one placement group by ID.

PlacementGroupInsertCall
Inserts a new placement group.

PlacementGroupListCall
Retrieves a list of placement groups, possibly filtered. This method supports paging.

PlacementGroupMethods
A builder providing access to all methods supported on placementGroup resources. It is not used directly, but through the Dfareporting hub.

PlacementGroupPatchCall
Updates an existing placement group. This method supports patch semantics.

PlacementGroupUpdateCall
Updates an existing placement group.

PlacementGroupsListResponse
Placement Group List Response

PlacementInsertCall
Inserts a new placement.

PlacementListCall
Retrieves a list of placements, possibly filtered. This method supports paging.

PlacementMethods
A builder providing access to all methods supported on placement resources. It is not used directly, but through the Dfareporting hub.

PlacementPatchCall
Updates an existing placement. This method supports patch semantics.

PlacementStrategiesListResponse
Placement Strategy List Response

PlacementStrategy
Contains properties of a placement strategy.

PlacementStrategyDeleteCall
Deletes an existing placement strategy.

PlacementStrategyGetCall
Gets one placement strategy by ID.

PlacementStrategyInsertCall
Inserts a new placement strategy.

PlacementStrategyListCall
Retrieves a list of placement strategies, possibly filtered. This method supports paging.

PlacementStrategyMethods
A builder providing access to all methods supported on placementStrategy resources. It is not used directly, but through the Dfareporting hub.

PlacementStrategyPatchCall
Updates an existing placement strategy. This method supports patch semantics.

PlacementStrategyUpdateCall
Updates an existing placement strategy.

PlacementTag
Placement Tag

PlacementUpdateCall
Updates an existing placement.

PlacementsGenerateTagsResponse
Placement GenerateTags Response

PlacementsListResponse
Placement List Response

PlatformType
Contains information about a platform type that can be targeted by ads.

PlatformTypeGetCall
Gets one platform type by ID.

PlatformTypeListCall
Retrieves a list of platform types.

PlatformTypeMethods
A builder providing access to all methods supported on platformType resources. It is not used directly, but through the Dfareporting hub.

PlatformTypesListResponse
Platform Type List Response

PopupWindowProperties
Popup Window Properties.

PostalCode
Contains information about a postal code that can be targeted by ads.

PostalCodeGetCall
Gets one postal code by ID.

PostalCodeListCall
Retrieves a list of postal codes.

PostalCodeMethods
A builder providing access to all methods supported on postalCode resources. It is not used directly, but through the Dfareporting hub.

PostalCodesListResponse
Postal Code List Response

Pricing
Pricing Information

PricingSchedule
Pricing Schedule

PricingSchedulePricingPeriod
Pricing Period

Project
Contains properties of a Planning project.

ProjectGetCall
Gets one project by ID.

ProjectListCall
Retrieves a list of projects, possibly filtered. This method supports paging .

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Dfareporting hub.

ProjectsListResponse
Project List Response

ReachReportCompatibleFields
Represents fields that are compatible to be selected for a report of type "REACH".

Recipient
Represents a recipient.

Region
Contains information about a region that can be targeted by ads.

RegionListCall
Retrieves a list of regions.

RegionMethods
A builder providing access to all methods supported on region resources. It is not used directly, but through the Dfareporting hub.

RegionsListResponse
Region List Response

RemarketingList
Contains properties of a remarketing list. Remarketing enables you to create lists of users who have performed specific actions on a site, then target ads to members of those lists. This resource can be used to manage remarketing lists that are owned by your advertisers. To see all remarketing lists that are visible to your advertisers, including those that are shared to your advertiser or account, use the TargetableRemarketingLists resource.

RemarketingListGetCall
Gets one remarketing list by ID.

RemarketingListInsertCall
Inserts a new remarketing list.

RemarketingListListCall
Retrieves a list of remarketing lists, possibly filtered. This method supports paging.

RemarketingListMethods
A builder providing access to all methods supported on remarketingList resources. It is not used directly, but through the Dfareporting hub.

RemarketingListPatchCall
Updates an existing remarketing list. This method supports patch semantics.

RemarketingListShare
Contains properties of a remarketing list's sharing information. Sharing allows other accounts or advertisers to target to your remarketing lists. This resource can be used to manage remarketing list sharing to other accounts and advertisers.

RemarketingListShareGetCall
Gets one remarketing list share by remarketing list ID.

RemarketingListShareMethods
A builder providing access to all methods supported on remarketingListShare resources. It is not used directly, but through the Dfareporting hub.

RemarketingListSharePatchCall
Updates an existing remarketing list share. This method supports patch semantics.

RemarketingListShareUpdateCall
Updates an existing remarketing list share.

RemarketingListUpdateCall
Updates an existing remarketing list.

RemarketingListsListResponse
Remarketing list response

Report
Represents a Report resource.

ReportCompatibleFieldQueryCall
Returns the fields that are compatible to be selected in the respective sections of a report criteria, given the fields already selected in the input report and user permissions.

ReportCompatibleFields
Represents fields that are compatible to be selected for a report of type "STANDARD".

ReportCriteria
The report criteria for a report of type "STANDARD".

ReportCrossDimensionReachCriteria
The report criteria for a report of type "CROSS_DIMENSION_REACH".

ReportDeleteCall
Deletes a report by its ID.

ReportDelivery
The report's email delivery settings.

ReportFileGetCall
Retrieves a report file by its report ID and file ID. This method supports media download.

ReportFileListCall
Lists files for a report.

ReportFloodlightCriteria
The report criteria for a report of type "FLOODLIGHT".

ReportFloodlightCriteriaReportProperties
The properties of the report.

ReportGetCall
Retrieves a report by its ID.

ReportInsertCall
Creates a report.

ReportList
Represents the list of reports.

ReportListCall
Retrieves list of reports.

ReportMethods
A builder providing access to all methods supported on report resources. It is not used directly, but through the Dfareporting hub.

ReportPatchCall
Updates an existing report. This method supports patch semantics.

ReportPathAttributionCriteria
The report criteria for a report of type "PATH_ATTRIBUTION".

ReportPathCriteria
The report criteria for a report of type "PATH".

ReportPathToConversionCriteria
The report criteria for a report of type "PATH_TO_CONVERSION".

ReportPathToConversionCriteriaReportProperties
The properties of the report.

ReportReachCriteria
The report criteria for a report of type "REACH".

ReportRunCall
Runs a report.

ReportSchedule
The report's schedule. Can only be set if the report's 'dateRange' is a relative date range and the relative date range is not "TODAY".

ReportUpdateCall
Updates a report.

ReportsConfiguration
Reporting Configuration

RichMediaExitOverride
Rich Media Exit Override.

Rule
A rule associates an asset with a targeting template for asset-level targeting. Applicable to INSTREAM_VIDEO creatives.

Site
Contains properties of a site.

SiteCompanionSetting
Companion Settings

SiteContact
Site Contact

SiteGetCall
Gets one site by ID.

SiteInsertCall
Inserts a new site.

SiteListCall
Retrieves a list of sites, possibly filtered. This method supports paging.

SiteMethods
A builder providing access to all methods supported on site resources. It is not used directly, but through the Dfareporting hub.

SitePatchCall
Updates an existing site. This method supports patch semantics.

SiteSettings
Site Settings

SiteSkippableSetting
Skippable Settings

SiteTranscodeSetting
Transcode Settings

SiteUpdateCall
Updates an existing site.

SiteVideoSettings
Video Settings

SitesListResponse
Site List Response

Size
Represents the dimensions of ads, placements, creatives, or creative assets.

SizeGetCall
Gets one size by ID.

SizeInsertCall
Inserts a new size.

SizeListCall
Retrieves a list of sizes, possibly filtered. Retrieved sizes are globally unique and may include values not currently in use by your account. Due to this, the list of sizes returned by this method may differ from the list seen in the Trafficking UI.

SizeMethods
A builder providing access to all methods supported on size resources. It is not used directly, but through the Dfareporting hub.

SizesListResponse
Size List Response

SkippableSetting
Skippable Settings

SortedDimension
Represents a sorted dimension.

Subaccount
Contains properties of a Campaign Manager subaccount.

SubaccountGetCall
Gets one subaccount by ID.

SubaccountInsertCall
Inserts a new subaccount.

SubaccountListCall
Gets a list of subaccounts, possibly filtered. This method supports paging.

SubaccountMethods
A builder providing access to all methods supported on subaccount resources. It is not used directly, but through the Dfareporting hub.

SubaccountPatchCall
Updates an existing subaccount. This method supports patch semantics.

SubaccountUpdateCall
Updates an existing subaccount.

SubaccountsListResponse
Subaccount List Response

TagData
Placement Tag Data

TagSetting
Tag Settings

TagSettings
Dynamic and Image Tag Settings.

TargetWindow
Target Window.

TargetableRemarketingList
Contains properties of a targetable remarketing list. Remarketing enables you to create lists of users who have performed specific actions on a site, then target ads to members of those lists. This resource is a read-only view of a remarketing list to be used to faciliate targeting ads to specific lists. Remarketing lists that are owned by your advertisers and those that are shared to your advertisers or account are accessible via this resource. To manage remarketing lists that are owned by your advertisers, use the RemarketingLists resource.

TargetableRemarketingListGetCall
Gets one remarketing list by ID.

TargetableRemarketingListListCall
Retrieves a list of targetable remarketing lists, possibly filtered. This method supports paging.

TargetableRemarketingListMethods
A builder providing access to all methods supported on targetableRemarketingList resources. It is not used directly, but through the Dfareporting hub.

TargetableRemarketingListsListResponse
Targetable remarketing list response

TargetingTemplate
Contains properties of a targeting template. A targeting template encapsulates targeting information which can be reused across multiple ads.

TargetingTemplateGetCall
Gets one targeting template by ID.

TargetingTemplateInsertCall
Inserts a new targeting template.

TargetingTemplateListCall
Retrieves a list of targeting templates, optionally filtered. This method supports paging.

TargetingTemplateMethods
A builder providing access to all methods supported on targetingTemplate resources. It is not used directly, but through the Dfareporting hub.

TargetingTemplatePatchCall
Updates an existing targeting template. This method supports patch semantics.

TargetingTemplateUpdateCall
Updates an existing targeting template.

TargetingTemplatesListResponse
Targeting Template List Response

TechnologyTargeting
Technology Targeting.

ThirdPartyAuthenticationToken
Third Party Authentication Token

ThirdPartyTrackingUrl
Third-party Tracking URL.

TranscodeSetting
Transcode Settings

UniversalAdId
A Universal Ad ID as per the VAST 4.0 spec. Applicable to the following creative types: INSTREAM_AUDIO, INSTREAM_VIDEO and VPAID.

UserDefinedVariableConfiguration
User Defined Variable configuration.

UserProfile
A UserProfile resource lets you list all DFA user profiles that are associated with a Google user account. The profile_id needs to be specified in other API requests.

UserProfileGetCall
Gets one user profile by ID.

UserProfileList
Represents the list of user profiles.

UserProfileListCall
Retrieves list of user profiles for a user.

UserProfileMethods
A builder providing access to all methods supported on userProfile resources. It is not used directly, but through the Dfareporting hub.

UserRole
Contains properties of auser role, which is used to manage user access.

UserRoleDeleteCall
Deletes an existing user role.

UserRoleGetCall
Gets one user role by ID.

UserRoleInsertCall
Inserts a new user role.

UserRoleListCall
Retrieves a list of user roles, possibly filtered. This method supports paging.

UserRoleMethods
A builder providing access to all methods supported on userRole resources. It is not used directly, but through the Dfareporting hub.

UserRolePatchCall
Updates an existing user role. This method supports patch semantics.

UserRolePermission
Contains properties of a user role permission.

UserRolePermissionGetCall
Gets one user role permission by ID.

UserRolePermissionGroup
Represents a grouping of related user role permissions.

UserRolePermissionGroupGetCall
Gets one user role permission group by ID.

UserRolePermissionGroupListCall
Gets a list of all supported user role permission groups.

UserRolePermissionGroupMethods
A builder providing access to all methods supported on userRolePermissionGroup resources. It is not used directly, but through the Dfareporting hub.

UserRolePermissionGroupsListResponse
User Role Permission Group List Response

UserRolePermissionListCall
Gets a list of user role permissions, possibly filtered.

UserRolePermissionMethods
A builder providing access to all methods supported on userRolePermission resources. It is not used directly, but through the Dfareporting hub.

UserRolePermissionsListResponse
User Role Permission List Response

UserRoleUpdateCall
Updates an existing user role.

UserRolesListResponse
User Role List Response

VideoFormat
Contains information about supported video formats.

VideoFormatGetCall
Gets one video format by ID.

VideoFormatListCall
Lists available video formats.

VideoFormatMethods
A builder providing access to all methods supported on videoFormat resources. It is not used directly, but through the Dfareporting hub.

VideoFormatsListResponse
Video Format List Response

VideoOffset
Video Offset

VideoSettings
Video Settings

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.