Struct google_dfareporting3d4::api::Conversion

pub struct Conversion {
    pub child_directed_treatment: Option<bool>,
    pub custom_variables: Option<Vec<CustomFloodlightVariable>>,
    pub dclid: Option<String>,
    pub encrypted_user_id: Option<String>,
    pub encrypted_user_id_candidates: Option<Vec<String>>,
    pub floodlight_activity_id: Option<String>,
    pub floodlight_configuration_id: Option<String>,
    pub gclid: Option<String>,
    pub kind: Option<String>,
    pub limit_ad_tracking: Option<bool>,
    pub match_id: Option<String>,
    pub mobile_device_id: Option<String>,
    pub non_personalized_ad: Option<bool>,
    pub ordinal: Option<String>,
    pub quantity: Option<String>,
    pub timestamp_micros: Option<String>,
    pub treatment_for_underage: Option<bool>,
    pub value: Option<f64>,
}

A Conversion represents when a user successfully performs a desired action after seeing an ad.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

batchinsert conversions (none)

batchupdate conversions (none)

Fields
child_directed_treatment: Option<bool>

Whether this particular request may come from a user under the age of 13, under COPPA compliance.

custom_variables: Option<Vec<CustomFloodlightVariable>>

Custom floodlight variables.

dclid: Option<String>

The display click ID. This field is mutually exclusive with encryptedUserId, encryptedUserIdCandidates[], matchId, mobileDeviceId and gclid. This or encryptedUserId or encryptedUserIdCandidates[] or matchId or mobileDeviceId or gclid is a required field.

encrypted_user_id: Option<String>

The alphanumeric encrypted user ID. When set, encryptionInfo should also be specified. This field is mutually exclusive with encryptedUserIdCandidates[], matchId, mobileDeviceId, gclid and dclid. This or encryptedUserIdCandidates[] or matchId or mobileDeviceId or gclid or dclid is a required field.

encrypted_user_id_candidates: Option<Vec<String>>

A list of the alphanumeric encrypted user IDs. Any user ID with exposure prior to the conversion timestamp will be used in the inserted conversion. If no such user ID is found then the conversion will be rejected with INVALID_ARGUMENT error. When set, encryptionInfo should also be specified. This field may only be used when calling batchinsert; it is not supported by batchupdate. This field is mutually exclusive with encryptedUserId, matchId, mobileDeviceId, gclid and dclid. This or encryptedUserId or matchId or mobileDeviceId or gclid or dclid is a required field.

floodlight_activity_id: Option<String>

Floodlight Activity ID of this conversion. This is a required field.

floodlight_configuration_id: Option<String>

Floodlight Configuration ID of this conversion. This is a required field.

gclid: Option<String>

The Google click ID. This field is mutually exclusive with encryptedUserId, encryptedUserIdCandidates[], matchId, mobileDeviceId and dclid. This or encryptedUserId or encryptedUserIdCandidates[] or matchId or mobileDeviceId or dclid is a required field.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#conversion".

limit_ad_tracking: Option<bool>

Whether Limit Ad Tracking is enabled. When set to true, the conversion will be used for reporting but not targeting. This will prevent remarketing.

match_id: Option<String>

The match ID field. A match ID is your own first-party identifier that has been synced with Google using the match ID feature in Floodlight. This field is mutually exclusive with encryptedUserId, encryptedUserIdCandidates[],mobileDeviceId, gclid and dclid. This or encryptedUserId or encryptedUserIdCandidates[] or mobileDeviceId or gclid or dclid is a required field.

mobile_device_id: Option<String>

The mobile device ID. This field is mutually exclusive with encryptedUserId, encryptedUserIdCandidates[], matchId, gclid and dclid. This or encryptedUserId or encryptedUserIdCandidates[] or matchId or gclid or dclid is a required field.

non_personalized_ad: Option<bool>

Whether the conversion was for a non personalized ad.

ordinal: Option<String>

The ordinal of the conversion. Use this field to control how conversions of the same user and day are de-duplicated. This is a required field.

quantity: Option<String>

The quantity of the conversion.

timestamp_micros: Option<String>

The timestamp of conversion, in Unix epoch micros. This is a required field.

treatment_for_underage: Option<bool>

Whether this particular request may come from a user under the age of 16 (may differ by country), under compliance with the European Union's General Data Protection Regulation (GDPR).

value: Option<f64>

The value of the conversion.