Struct google_dfareporting3d4::api::Placement

pub struct Placement {
    pub account_id: Option<String>,
    pub ad_blocking_opt_out: Option<bool>,
    pub additional_sizes: Option<Vec<Size>>,
    pub advertiser_id: Option<String>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub archived: Option<bool>,
    pub campaign_id: Option<String>,
    pub campaign_id_dimension_value: Option<DimensionValue>,
    pub comment: Option<String>,
    pub compatibility: Option<String>,
    pub content_category_id: Option<String>,
    pub create_info: Option<LastModifiedInfo>,
    pub directory_site_id: Option<String>,
    pub directory_site_id_dimension_value: Option<DimensionValue>,
    pub external_id: Option<String>,
    pub id: Option<String>,
    pub id_dimension_value: Option<DimensionValue>,
    pub key_name: Option<String>,
    pub kind: Option<String>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub lookback_configuration: Option<LookbackConfiguration>,
    pub name: Option<String>,
    pub payment_approved: Option<bool>,
    pub payment_source: Option<String>,
    pub placement_group_id: Option<String>,
    pub placement_group_id_dimension_value: Option<DimensionValue>,
    pub placement_strategy_id: Option<String>,
    pub pricing_schedule: Option<PricingSchedule>,
    pub primary: Option<bool>,
    pub publisher_update_info: Option<LastModifiedInfo>,
    pub site_id: Option<String>,
    pub site_id_dimension_value: Option<DimensionValue>,
    pub size: Option<Size>,
    pub ssl_required: Option<bool>,
    pub status: Option<String>,
    pub subaccount_id: Option<String>,
    pub tag_formats: Option<Vec<String>>,
    pub tag_setting: Option<TagSetting>,
    pub video_active_view_opt_out: Option<bool>,
    pub video_settings: Option<VideoSettings>,
    pub vpaid_adapter_choice: Option<String>,
}

Contains properties of a placement.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

generatetags placements (none)

get placements (response)

insert placements (request|response)

list placements (none)

patch placements (request|response)

update placements (request|response)

Fields
account_id: Option<String>

Account ID of this placement. This field can be left blank.

ad_blocking_opt_out: Option<bool>

Whether this placement opts out of ad blocking. When true, ad blocking is disabled for this placement. When false, the campaign and site settings take effect.

additional_sizes: Option<Vec<Size>>

Additional sizes associated with this placement. When inserting or updating a placement, only the size ID field is used.

advertiser_id: Option<String>

Advertiser ID of this placement. This field can be left blank.

advertiser_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the advertiser. This is a read-only, auto-generated field.

archived: Option<bool>

Whether this placement is archived.

campaign_id: Option<String>

Campaign ID of this placement. This field is a required field on insertion.

campaign_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the campaign. This is a read-only, auto-generated field.

comment: Option<String>

Comments for this placement.

compatibility: Option<String>

Placement compatibility. DISPLAY and DISPLAY_INTERSTITIAL refer to rendering on desktop, on mobile devices or in mobile apps for regular or interstitial ads respectively. APP and APP_INTERSTITIAL are no longer allowed for new placement insertions. Instead, use DISPLAY or DISPLAY_INTERSTITIAL. IN_STREAM_VIDEO refers to rendering in in-stream video ads developed with the VAST standard. This field is required on insertion.

content_category_id: Option<String>

ID of the content category assigned to this placement.

create_info: Option<LastModifiedInfo>

Information about the creation of this placement. This is a read-only field.

directory_site_id: Option<String>

Directory site ID of this placement. On insert, you must set either this field or the siteId field to specify the site associated with this placement. This is a required field that is read-only after insertion.

directory_site_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the directory site. This is a read-only, auto-generated field.

external_id: Option<String>

External ID for this placement.

id: Option<String>

ID of this placement. This is a read-only, auto-generated field.

id_dimension_value: Option<DimensionValue>

Dimension value for the ID of this placement. This is a read-only, auto-generated field.

key_name: Option<String>

Key name of this placement. This is a read-only, auto-generated field.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#placement".

last_modified_info: Option<LastModifiedInfo>

Information about the most recent modification of this placement. This is a read-only field.

lookback_configuration: Option<LookbackConfiguration>

Lookback window settings for this placement.

name: Option<String>

Name of this placement.This is a required field and must be less than or equal to 256 characters long.

payment_approved: Option<bool>

Whether payment was approved for this placement. This is a read-only field relevant only to publisher-paid placements.

payment_source: Option<String>

Payment source for this placement. This is a required field that is read-only after insertion.

placement_group_id: Option<String>

ID of this placement's group, if applicable.

placement_group_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the placement group. This is a read-only, auto-generated field.

placement_strategy_id: Option<String>

ID of the placement strategy assigned to this placement.

pricing_schedule: Option<PricingSchedule>

Pricing schedule of this placement. This field is required on insertion, specifically subfields startDate, endDate and pricingType.

primary: Option<bool>

Whether this placement is the primary placement of a roadblock (placement group). You cannot change this field from true to false. Setting this field to true will automatically set the primary field on the original primary placement of the roadblock to false, and it will automatically set the roadblock's primaryPlacementId field to the ID of this placement.

publisher_update_info: Option<LastModifiedInfo>

Information about the last publisher update. This is a read-only field.

site_id: Option<String>

Site ID associated with this placement. On insert, you must set either this field or the directorySiteId field to specify the site associated with this placement. This is a required field that is read-only after insertion.

site_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the site. This is a read-only, auto-generated field.

size: Option<Size>

Size associated with this placement. When inserting or updating a placement, only the size ID field is used. This field is required on insertion.

ssl_required: Option<bool>

Whether creatives assigned to this placement must be SSL-compliant.

status: Option<String>

Third-party placement status.

subaccount_id: Option<String>

Subaccount ID of this placement. This field can be left blank.

tag_formats: Option<Vec<String>>

Tag formats to generate for this placement. This field is required on insertion. Acceptable values are: - "PLACEMENT_TAG_STANDARD" - "PLACEMENT_TAG_IFRAME_JAVASCRIPT" - "PLACEMENT_TAG_IFRAME_ILAYER" - "PLACEMENT_TAG_INTERNAL_REDIRECT" - "PLACEMENT_TAG_JAVASCRIPT" - "PLACEMENT_TAG_INTERSTITIAL_IFRAME_JAVASCRIPT" - "PLACEMENT_TAG_INTERSTITIAL_INTERNAL_REDIRECT" - "PLACEMENT_TAG_INTERSTITIAL_JAVASCRIPT" - "PLACEMENT_TAG_CLICK_COMMANDS" - "PLACEMENT_TAG_INSTREAM_VIDEO_PREFETCH" - "PLACEMENT_TAG_INSTREAM_VIDEO_PREFETCH_VAST_3" - "PLACEMENT_TAG_INSTREAM_VIDEO_PREFETCH_VAST_4" - "PLACEMENT_TAG_TRACKING" - "PLACEMENT_TAG_TRACKING_IFRAME" - "PLACEMENT_TAG_TRACKING_JAVASCRIPT"

tag_setting: Option<TagSetting>

Tag settings for this placement.

video_active_view_opt_out: Option<bool>

Whether Verification and ActiveView are disabled for in-stream video creatives for this placement. The same setting videoActiveViewOptOut exists on the site level -- the opt out occurs if either of these settings are true. These settings are distinct from DirectorySites.settings.activeViewOptOut or Sites.siteSettings.activeViewOptOut which only apply to display ads. However, Accounts.activeViewOptOut opts out both video traffic, as well as display ads, from Verification and ActiveView.

video_settings: Option<VideoSettings>

A collection of settings which affect video creatives served through this placement. Applicable to placements with IN_STREAM_VIDEO compatibility.

vpaid_adapter_choice: Option<String>

VPAID adapter setting for this placement. Controls which VPAID format the measurement adapter will use for in-stream video creatives assigned to this placement. Note: Flash is no longer supported. This field now defaults to HTML5 when the following values are provided: FLASH, BOTH.